+++ src/components/Notifications.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import defaultMessage, {
   fetchMessage,
 } from "../data/notifications.messages.data";
@@ -6,6 +6,7 @@ import "../styles/Notifications.css";
 
 export default function Notifications({ currentScore }) {
   const [message, setMessage] = useState(null);
+  const defaultRef = useRef(null);
 
   useEffect(() => {
     if (currentScore > 0) {
@@ -13,12 +14,17 @@ export default function Notifications({ currentScore }) {
     } else {
       setMessage(defaultMessage);
     }
+
+    defaultRef?.current.classList.add("slide");
+    setTimeout(() => {
+      defaultRef?.current.classList.remove("slide");
+    }, 350);
   }, [currentScore]);
 
   return (
-    <div id="notifications">
+    <div id="notifications" role="alert">
       <div className="messages">
-        <p>{message}</p>
+        <p ref={defaultRef}>{message}</p>
       </div>
     </div>
   );